requirements.txt
aiofiles==24.1.0
aiohttp==3.9.5
aiosignal==1.3.1
alabaster==0.7.16
altair==5.3.0
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
astroid==3.2.4
async-timeout==4.0.3
attrs==23.2.0
Authlib==1.6.0
Babel==2.15.0
beautifulsoup4==4.13.4
blinker==1.8.2
cachetools==5.3.3
certifi==2024.2.2
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==45.0.4
dataclasses-json==0.6.6
dill==0.3.8
distro==1.9.0
Django==5.1.3
django-cors-headers==4.6.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
docutils==0.21.2
esprima==4.0.1
et-xmlfile==1.1.0
exceptiongroup==1.3.0
faiss-cpu==1.11.0
fastapi==0.115.12
fastmcp==2.8.1
filelock==3.15.4
frozenlist==1.4.1
fsspec==2024.6.1
gitdb==4.0.11
GitPython==3.1.43
google-ai-generativelanguage==0.6.4
google-api-core==2.19.0
google-api-python-client==2.129.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-cloud==0.34.0
google-generativeai==0.5.4
googleapis-common-protos==1.63.0
greenlet==3.0.3
grpcio==1.64.0
grpcio-status==1.62.2
h11==0.14.0
html5lib==1.1
httpcore==1.0.5
httplib2==0.22.0
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.24.5
idna==3.7
imagesize==1.4.1
importlib-metadata==7.0.1
isort==5.13.2
javalang==0.13.0
Jinja2==3.1.4
joblib==1.5.1
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
langchain==0.2.0
langchain-community==0.2.0
langchain-core==0.2.1
langchain-google-genai==1.0.5
langchain-openai==0.1.7
langchain-text-splitters==0.2.0
langsmith==0.1.62
lxml==5.4.0
markdown-it-py==3.0.0
markdownify==1.1.0
MarkupSafe==2.1.5
marshmallow==3.21.2
mccabe==0.7.0
mcp==1.9.4
mdurl==0.1.2
mpmath==1.3.0
multidict==6.0.5
mypy-extensions==1.0.0
mysql-connector-python==8.4.0
networkx==3.4.2
numpy==1.26.4
openai==1.30.1
openapi-pydantic==0.5.1
openpyxl==3.1.2
orjson==3.10.3
packaging==23.2
pandas==2.2.2
pdfminer.six==20250506
pdfplumber==0.11.7
pillow==10.3.0
pip-autoremove==0.10.0
plantuml==0.3.0
platformdirs==4.2.2
proto-plus==1.23.0
protobuf==4.25.3
psycopg2-binary==2.9.9
py4j==0.10.9.7
pyarrow==16.1.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycallgraph2==1.1.3
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.9.1
pydantic_core==2.33.2
pydeck==0.9.1
pydot==3.0.2
pydub==0.25.1
Pygments==2.18.0
PyJWT==2.10.1
pylint==3.2.6
PyMuPDF==1.26.0
pyparsing==3.1.2
PyPDF2==3.0.1
pypdfium2==4.30.1
pyspark==3.5.1
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.1.0
python-multipart==0.0.20
pytz==2024.1
PyYAML==6.0.1
readabilipy==0.3.0
redis==6.2.0
referencing==0.35.1
regex==2024.5.15
reportlab==4.1.0
requests==2.32.2
rich==14.0.0
rpds-py==0.18.1
rsa==4.9
safetensors==0.4.4
scikit-learn==1.7.0
scipy==1.15.3
sentence-transformers==4.1.0
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
snowballstemmer==2.2.0
soupsieve==2.7
SpeechRecognition==3.10.4
Sphinx==7.4.7
sphinx-autodoc-typehints==2.2.3
sphinxcontrib-applehelp==1.0.8
sphinxcontrib-devhelp==1.0.6
sphinxcontrib-htmlhelp==2.0.6
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.8
sphinxcontrib-serializinghtml==1.1.10
SQLAlchemy==2.0.30
sqlparse==0.5.2
sse-starlette==2.3.6
starlette==0.46.2
streamlit==1.34.0
sympy==1.14.0
tenacity==8.3.0
threadpoolctl==3.6.0
tiktoken==0.7.0
tokenizers==0.19.1
toml==0.10.2
tomli==2.0.1
tomlkit==0.13.0
toolz==0.12.1
torch==2.7.1
tornado==6.4
tqdm==4.66.4
transformers==4.44.0
typer==0.16.0
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2024.1
tzlocal==5.2
uritemplate==4.1.1
urllib3==2.2.1
uvicorn==0.34.2
validate-email==1.3
validators==0.22.0
watchdog==4.0.0
webencodings==0.5.1
yarl==1.9.4
zipp==3.17.0
